+++ rust/processor/src/db/common/models/emojicoin_models/models/arena_enter_event.rs
@@ -2,6 +2,7 @@ use crate::{
     db::common::models::emojicoin_models::json_types::{ArenaEnterEvent, TxnInfo},
     schema::arena_enter_events,
 };
+use bigdecimal::BigDecimal;
 use field_count::FieldCount;
 use serde::{Deserialize, Serialize};
 
@@ -17,18 +18,18 @@ pub struct ArenaEnterEventModel {
     pub transaction_timestamp: chrono::NaiveDateTime,
 
     pub user: String,
-    pub melee_id: i64,
-    pub input_amount: i64,
-    pub quote_volume: i64,
-    pub integrator_fee: i64,
-    pub match_amount: i64,
+    pub melee_id: BigDecimal,
+    pub input_amount: BigDecimal,
+    pub quote_volume: BigDecimal,
+    pub integrator_fee: BigDecimal,
+    pub match_amount: BigDecimal,
 
-    pub emojicoin_0_proceeds: i64,
-    pub emojicoin_1_proceeds: i64,
-    pub emojicoin_0_exchange_rate_base: i64,
-    pub emojicoin_0_exchange_rate_quote: i64,
-    pub emojicoin_1_exchange_rate_base: i64,
-    pub emojicoin_1_exchange_rate_quote: i64,
+    pub emojicoin_0_proceeds: BigDecimal,
+    pub emojicoin_1_proceeds: BigDecimal,
+    pub emojicoin_0_exchange_rate_base: BigDecimal,
+    pub emojicoin_0_exchange_rate_quote: BigDecimal,
+    pub emojicoin_1_exchange_rate_base: BigDecimal,
+    pub emojicoin_1_exchange_rate_quote: BigDecimal,
 }
 
 impl ArenaEnterEventModel {
